+++ ddth-cache-adapter-core/src/main/java/com/github/ddth/cacheadapter/ces/JbossCacheEntrySerializer.java
@@ -12,6 +12,7 @@ import com.github.ddth.commons.utils.SerializationUtils;
  * 
  * @author Thanh Nguyen <btnguyen2k@gmail.com>
  * @since 0.3.0
+ * @deprecated since 0.5.1
  */
 public class JbossCacheEntrySerializer extends AbstractCacheEntrySerializer {