+++ ddth-cache-adapter-core/src/test/java/com/github/ddth/cacheadapter/qnd/QndCacheSize.java
@@ -23,12 +23,33 @@ public class QndCacheSize {
 
         System.out.println("serDDTHCommons(raw): " + SerializationUtils.toByteArray(value).length);
         System.out.println(
+                "serJboss(raw)      : " + SerializationUtils.toByteArrayJboss(value).length);
+        System.out.println(
                 "serKyro(raw)       : " + SerializationUtils.toByteArrayKryo(value).length);
+        System.out
+                .println("serFst(raw)        : " + SerializationUtils.toByteArrayFst(value).length);
         System.out.println("serDDTHCommons(ce) : " + SerializationUtils.toByteArray(ce).length);
+        System.out
+                .println("serJboss(ce)       : " + SerializationUtils.toByteArrayJboss(ce).length);
         System.out.println("serKyro(ce)        : " + SerializationUtils.toByteArrayKryo(ce).length);
+        System.out.println("serFst(ce)         : " + SerializationUtils.toByteArrayFst(ce).length);
         System.out.println("serThrift(ce)      : " + ThriftUtils.serialize(ce).length);
 
-        System.out.println("==== Ser/Deser:");
+        System.out.println("==== Ser/Deser (Raw):");
+        {
+            byte[] dataRaw = SerializationUtils.toByteArray(value);
+            System.out.println("Size (Deft/raw) : " + dataRaw.length);
+            Object obj = SerializationUtils.fromByteArray(dataRaw, value.getClass());
+            System.out.println("Deserialize     : " + obj);
+            System.out.println("Compare         : ===>" + compare(value, obj));
+        }
+        {
+            byte[] dataRaw = SerializationUtils.toByteArrayJboss(value);
+            System.out.println("Size (Jboss/raw): " + dataRaw.length);
+            Object obj = SerializationUtils.fromByteArrayJboss(dataRaw, value.getClass());
+            System.out.println("Deserialize     : " + obj);
+            System.out.println("Compare         : ===>" + compare(value, obj));
+        }
         {
             byte[] dataRaw = SerializationUtils.toByteArrayKryo(value);
             System.out.println("Size (Kryo/raw) : " + dataRaw.length);
@@ -37,9 +58,25 @@ public class QndCacheSize {
             System.out.println("Compare         : ===>" + compare(value, obj));
         }
         {
+            byte[] dataRaw = SerializationUtils.toByteArrayFst(value);
+            System.out.println("Size (Fst/raw)  : " + dataRaw.length);
+            Object obj = SerializationUtils.fromByteArrayFst(dataRaw, value.getClass());
+            System.out.println("Deserialize     : " + obj);
+            System.out.println("Compare         : ===>" + compare(value, obj));
+        }
+
+        System.out.println("==== Ser/Deser (CE):");
+        {
             byte[] dataCe = SerializationUtils.toByteArray(ce);
-            System.out.println("Size (Jboss/ce) : " + dataCe.length);
+            System.out.println("Size (Deft/ce)  : " + dataCe.length);
             Object obj = SerializationUtils.fromByteArray(dataCe, CacheEntry.class);
+            System.out.println("Deserialize     : " + obj);
+            System.out.println("Compare         : ===>" + compare(ce, obj));
+        }
+        {
+            byte[] dataCe = SerializationUtils.toByteArrayJboss(ce);
+            System.out.println("Size (Jboss/ce) : " + dataCe.length);
+            Object obj = SerializationUtils.fromByteArrayJboss(dataCe, CacheEntry.class);
             System.out.println("Deserialize     : " + obj);
             System.out.println("Compare         : ===>" + compare(ce, obj));
         }
@@ -47,6 +84,13 @@ public class QndCacheSize {
             byte[] dataCe = SerializationUtils.toByteArrayKryo(ce);
             System.out.println("Size (Kryo/ce)  : " + dataCe.length);
             Object obj = SerializationUtils.fromByteArrayKryo(dataCe, CacheEntry.class);
+            System.out.println("Deserialize     : " + obj);
+            System.out.println("Compare         : ===>" + compare(ce, obj));
+        }
+        {
+            byte[] dataCe = SerializationUtils.toByteArrayFst(ce);
+            System.out.println("Size (Fst/ce)   : " + dataCe.length);
+            Object obj = SerializationUtils.fromByteArrayFst(dataCe, CacheEntry.class);
             System.out.println("Deserialize     : " + obj);
             System.out.println("Compare         : ===>" + compare(ce, obj));
         }